Fix vertical re-positioning
Before we were tracking the current column so that when a vertical
motion would be performed we would move point back to that column.  The
idea is to make motions in a grid view intuitive.  However, the
Completions' buffer uses tabs to align candidates, so this approach
would fail to move to the candidate above/below if point would land in a
tabbed space.  What it would do instead is move vertically and then
shift to the next completion after point.  Whereas now if it lands in a
tab, it goes to the candidate before point.

This is an older bug, but we repurpose the code from d3aa0ca to take
care of it.