Refine switching to top/bottom of Completions
Now the candidate should be highlighted right away.  Before it would
fail to do so for a new Completions' buffer.

Also, now it works properly when completions-detailed is non-nil and we
are switching to the bottom of the results produced by
describe-symbol (or related).  Before it would go to the first character
which is just an annotation character.