Remove mct--region-buf
We do not have to store this buffer since completion-reference-buffer exists.
completion-in-region-mode is automatically left as soon as we leave the original
buffer. This happens in the completion-in-region--postch hook function.

Unfortunately this makes it hard to implement a command bound to TAB, which
inserts the current candidate and proceeds with the completion in the buffer
(mct-region-choose-completion-no-exit). It would be nice to have this for file
path completion. There are these possibilities to achieve this goal:

1) Remove the completion-in-region--postch from the post-command-hook list.
Instead use our own hook which preserves the completion-in-region-mode. This
approach is probably not such a good idea, since by definition of the
completion-in-region-mode, the mode should be left as soon as we leave the
buffer.

2) Save the entire completion-in-region--data just before entering the
*Completions* buffer. Use the saved data to restore the completion-in-region
after mct-region-choose-completion-no-exit, such

3) Try to restart the completion after mct-region-choose-completion-no-exit, by
calling completion-at-point again or the appropriate completion-in-region
command. I don't think it is easy to find a way which works in general.

I would try to implement option 2.