Remove task for mct-choose-completion-number
We cannot make it work in its current state for mct-region-mode.  The
reason is that as soon as the minibuffer is activated to select a line
number, the Completions' window gets removed if the point was not
already inside of it (so if it was in the buffer that started
completion-in-region).

I tried various functions to make the window dedicated, in the hope that
it would remain on display, but none worked.  Perhaps I did something
wrong though, so if someone has a better idea, I am happy to learn about
it.