Fix Emacs 29 regression for out-of-bounds motion
MCT accepts a numeric argument for motions inside the Completions
buffer.  The issue has been how to handle those cases where the ARGth
motion would exceed the boundaries.  For example, if we are on line 3
and have a -10 motion, the point will be placed in the minibuffer
instead of trying to cycle back to the minibuffer, then the bottom of
the Completion, and then move however many remaining places backward.

Just switching to the minibuffer when the ARG is out-of-bounds makes the
code easier and we save the user from the trouble of having to target
the minibuffer with precision (the last point in the Completions is
still relevant, such as when using mct-edit-completion).

Anyhow, Emacs 29 changes how the next-completion/previous-completion
work, where they now return an out-of-bounds error, so we have to make
some adjustments.  My tests suggest that this patch addresses the
underlying problem, while it has no regressions on Emacs 27.