BREAKING: Remove line number functionality
The original idea was to use line numbers the way we do in ordinary
buffers.  That works in most cases, as completion candidates tend to be
a few characters long.  However, there are instances where the candidate
is a longer stretch of text that includes newline characters, page
breaks, and the like: the assumption of one candidate per line number no
longer holds true, leading to an awkward experience.

Users are encouraged to give Avy a try.  We provide the mct-avy.el and
document how to use it in the manual.  In short, it targets a candidate
without moving the point to it.