Make the Completions persistent when visible or with passlist command
This means that when we hit TAB to complete the common prefix, we
still see the Completions afterwards.

The 'mct-persist-dynamic-completion' is a needless option that is
there for some implementation details. Users who want 'find-file' or
related to persist the Completions buffer should add that command to
the 'mct-completion-passlist'.

I am inspired to make this change in response to a question posed by
Ryan Davis regarding the behaviour of 'mct-persist-dynamic-completion'.
This was done in issue 7: <https://github.com/protesilaos/mct/issues/7>.