Depend on 29.4 instead of 29.1
Thanks to Ryan Davis for informing me that 29.1 does not define the
variable minibuffer-completion-base. This was done in issue 11:
<https://github.com/protesilaos/mct/issues/11>.

We could work around it, though it is better to bump the version we
depend on. This is because the higher number is the one that includes
the security patches that were publicised recently. Cannot find the
links right now...